+++ example/my_model.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 
-part 'typicode_model.g.dart';
+part 'my_model.g.dart';
 
 // We use the package json_serializable and json_annotation here to make it easier to serialize and deserialize json objects.
 // To generate the outputs, run: `dart run build_runner build --delete-conflicting-outputs`.
@@ -8,7 +8,7 @@ part 'typicode_model.g.dart';
 
 // // See: https://jsonplaceholder.typicode.com/ for more informations about the free fake API for testing and prototyping
 @JsonSerializable()
-class TypicodeModel {
+class MyModel {
   final int? userId;
   final int? id;
   final String? title;
@@ -17,26 +17,26 @@ class TypicodeModel {
   // Creates a new instance of an `TypicodeModel`.
   //
   // See: https://jsonplaceholder.typicode.com/ for more informations
-  TypicodeModel(this.userId, this.id, this.title, this.body);
+  MyModel(this.userId, this.id, this.title, this.body);
 
   // Creates a new instance of `HttpError` from a map.
   //
   // This factory constructor allows for the creation of an `HttpError`
   // instance from a JSON map. This is useful when receiving an HTTP
   // error in the form of JSON from an API.
-  factory TypicodeModel.fromJson(Map<String, dynamic> json) =>
-      _$TypicodeModelFromJson(json);
+  factory MyModel.fromJson(Map<String, dynamic> json) =>
+      _$MyModelFromJson(json);
 
   // Converts the `TypicodeModel` instance to a map.
   //
   // This method serializes the `TypicodeModel` instance into a JSON map,
   // which can be useful when sending HTTP errors to an API in the
   // form of JSON.
-  Map<String, dynamic> toJson() => _$TypicodeModelToJson(this);
+  Map<String, dynamic> toJson() => _$MyModelToJson(this);
 
   // Let's provide a toString method so we can debug it better.
   @override
   String toString() {
-    return 'TypicodeModel(\n  userId: $userId, \n  id: $id, \n  title: $title, \n  body: $body)';
+    return 'MyModel(\n  userId: $userId, \n  id: $id, \n  title: $title, \n  body: $body)';
   }
 }