+++ example/my_model.g.dart
@@ -1,21 +1,19 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'typicode_model_example.dart';
+part of 'my_model.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-TypicodeModel _$TypicodeModelFromJson(Map<String, dynamic> json) =>
-    TypicodeModel(
+MyModel _$MyModelFromJson(Map<String, dynamic> json) => MyModel(
       json['userId'] as int?,
       json['id'] as int?,
       json['title'] as String?,
       json['body'] as String?,
     );
 
-Map<String, dynamic> _$TypicodeModelToJson(TypicodeModel instance) =>
-    <String, dynamic>{
+Map<String, dynamic> _$MyModelToJson(MyModel instance) => <String, dynamic>{
       'userId': instance.userId,
       'id': instance.id,
       'title': instance.title,